More bugs but no fixes
Dear Mr.Chan,

This will be the final commit and pong 1% of the time doesnt work as intended but it works 99% of the time and I find that ok. I will never touch this game again. If Iven wants to fix bugs he can. Im moving on to the graphic user interface

Best regards,
VVarren